Display public_updated_at on dataset show page
We display a dataset's `public_updated_at` value on the dataset show page when
it's present. This will be the value that's shown permanently after we have
reindexed Elasticsearch to include it, so we will remove the rest of the logic
in the `displayed_date` helper after this has been done.